requirements.txt
flask
flask-socketio
eventlet
newspaper3k
nltk